# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Squeak Platform",
        "title": "Squeak API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.squeak.today",
    "basePath": "/",
    "paths": {
...
        "/profile": {
            "get": {
                "description": "Get the user's profile information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetProfileResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile/upsert": {
            "post": {
                "description": "Create or update the user's profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Upsert user profile",
                "parameters": [
                    {
                        "description": "Profile information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpsertProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpsertProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/progress/increment": {
            "get": {
                "description": "Increment the number of questions completed for today",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Increment questions completed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Amount to increment by",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IncrementProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/qna": {
            "post": {
                "description": "Get an existing question or generate a new one for the given content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qna"
                ],
                "summary": "Get or generate a question",
                "parameters": [
                    {
                        "description": "Question request parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetQuestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/qna/evaluate": {
            "post": {
                "description": "Evaluate a user's answer to a question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qna"
                ],
                "summary": "Evaluate an answer",
                "parameters": [
                    {
                        "description": "Answer evaluation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EvaluateAnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EvaluateAnswerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ERROR_CODE": {
            "type": "string",
            "enum": [
                "PROFILE_NOT_FOUND",
                "NO_TRANSCRIPT",
                "AUTH_REQUIRED",
                "USAGE_LIMIT_REACHED",
                "USAGE_RESTRICTED"
            ],
            "x-enum-varnames": [
                "PROFILE_NOT_FOUND",
                "NO_TRANSCRIPT",
                "AUTH_REQUIRED",
                "USAGE_LIMIT_REACHED",
                "USAGE_RESTRICTED"
            ]
        },
        "models.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "code": {
                    "enum": [
                        "PROFILE_NOT_FOUND",
                        "NO_TRANSCRIPT",
                        "AUTH_REQUIRED",
                        "USAGE_LIMIT_REACHED",
                        "USAGE_RESTRICTED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ERROR_CODE"
                        }
                    ],
                    "example": "PROFILE_NOT_FOUND"
                },
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
        "models.EvaluateAnswerRequest": {
            "type": "object",
            "required": [
                "answer",
                "cefr",
                "content",
                "question"
            ],
            "properties": {
                "answer": {
                    "type": "string",
                    "example": "Hello"
                },
                "cefr": {
                    "type": "string",
                    "example": "B1"
                },
                "content": {
                    "type": "string",
                    "example": "Bonjour, comment ça va?"
                },
                "question": {
                    "type": "string",
                    "example": "What does 'bonjour' mean?"
                }
            }
        },
        "models.EvaluateAnswerResponse": {
            "type": "object",
            "required": [
                "evaluation",
                "explanation"
            ],
            "properties": {
                "evaluation": {
                    "type": "string",
                    "example": "PASS"
                },
                "explanation": {
                    "type": "string",
                    "example": "Perfect!"
                }
            }
        },
...
        "models.GetProfileResponse": {
            "type": "object",
            "required": [
                "interested_topics",
                "learning_language",
                "skill_level",
                "username"
            ],
            "properties": {
                "daily_questions_goal": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3
                },
                "interested_topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"NBA\"]"
                    ]
                },
                "learning_language": {
                    "type": "string",
                    "example": "French"
                },
                "skill_level": {
                    "type": "string",
                    "example": "B1"
                },
                "username": {
                    "type": "string",
                    "example": "connortbot"
                }
            }
        },
        "models.GetQuestionRequest": {
            "type": "object",
            "required": [
                "cefr_level",
                "content_type",
                "id",
                "question_type"
            ],
            "properties": {
                "cefr_level": {
                    "type": "string",
                    "example": "B1"
                },
                "content_type": {
                    "type": "string",
                    "example": "News"
                },
                "id": {
                    "type": "string",
                    "example": "123"
                },
                "question_type": {
                    "type": "string",
                    "enum": [
                        "vocab",
                        "understanding"
                    ],
                    "example": "vocab"
                }
            }
        },
        "models.GetQuestionResponse": {
            "type": "object",
            "required": [
                "question"
            ],
            "properties": {
                "question": {
                    "type": "string",
                    "example": "What does 'bonjour' mean?"
                }
            }
        },
...
        "models.IncrementProgressResponse": {
            "type": "object",
            "required": [
                "date",
                "goal_met",
                "user_id"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2025-02-26T00:00:00Z"
                },
                "goal_met": {
                    "type": "boolean",
                    "example": true
                },
                "questions_completed": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 5
                },
                "user_id": {
                    "type": "string",
                    "example": "123"
                }
            }
        },
...
        "models.UpsertProfileRequest": {
            "type": "object",
            "required": [
                "interested_topics",
                "learning_language",
                "skill_level",
                "username"
            ],
            "properties": {
                "daily_questions_goal": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3
                },
                "interested_topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"NBA\"]"
                    ]
                },
                "learning_language": {
                    "type": "string",
                    "example": "French"
                },
                "skill_level": {
                    "type": "string",
                    "example": "B1"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "models.UpsertProfileResponse": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 123
                },
                "message": {
                    "type": "string",
                    "example": "Profile updated successfully"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT Authorization header using Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}